+++ tests/test_photon_array.py
@@ -18,6 +18,8 @@
 
 from __future__ import print_function
 import numpy as np
+import os
+import sys
 
 from galsim_test_helpers import *
 
@@ -177,7 +179,78 @@ def test_wavelength_sampler():
     print('sum = ',im1.array.sum(),im2.array.sum())
     np.testing.assert_array_equal(im1.array, im2.array)
 
+@timer
+def test_photon_angles():
+    """Test the photon_array function
+    """
+    # Make a photon array
+    seed = 12345
+    ud = galsim.UniformDeviate(seed)
+    gal = galsim.Sersic(n=4, half_light_radius=1)
+    photon_array = gal.SBProfile.shoot(100000, ud)
+
+    # Add the directions (N.B. using the same seed as for generating the photon array
+    # above.  The fact that it is the same does not matter here; the testing routine
+    # only needs to have a definite seed value so the consistency of the results with
+    # expectations can be evaluated precisely
+    fratio = 1.2
+    obscuration = 0.2
+
+    # rng can be None, an existing BaseDeviate, or an integer
+    for rng in [ None, ud, 12345 ]:
+        assigner = galsim.FRatioAngles(fratio, obscuration, rng)
+        assigner.applyTo(photon_array)
+
+        dxdz = photon_array.getDXDZArray()
+        dydz = photon_array.getDYDZArray()
+
+        phi = np.arctan2(dydz,dxdz)
+        tantheta = np.sqrt(np.square(dxdz) + np.square(dydz))
+        sintheta = np.sin(np.arctan(tantheta))
+
+        # Check that the values are within the ranges expected
+        # (The test on phi really can't fail, because it is only testing the range of the
+        # arctan2 function.)
+        np.testing.assert_array_less(-phi, np.pi, "Azimuth angles outside possible range")
+        np.testing.assert_array_less(phi, np.pi, "Azimuth angles outside possible range")
+
+        fov_angle = np.arctan(0.5 / fratio)
+        obscuration_angle = obscuration * fov_angle
+        np.testing.assert_array_less(-sintheta, -np.sin(obscuration_angle),
+                                     "Inclination angles outside possible range")
+        np.testing.assert_array_less(sintheta, np.sin(fov_angle),
+                                     "Inclination angles outside possible range")
+
+    # Compare these slopes with the expected distributions (uniform in azimuth 
+    # over all azimiths and uniform in sin(inclination) over the range of
+    # allowed inclinations
+    # Only test this for the last one, so we make sure we have a deterministic result.
+    # (The above tests should be reliable even for the default rng.)
+    phi_histo, phi_bins = np.histogram(phi, bins=100)
+    sintheta_histo, sintheta_bins = np.histogram(sintheta, bins=100)
+    phi_ref = float(np.sum(phi_histo))/phi_histo.size
+    sintheta_ref = float(np.sum(sintheta_histo))/sintheta_histo.size
+
+    chisqr_phi = np.sum(np.square(phi_histo - phi_ref)/phi_ref) / phi_histo.size
+    chisqr_sintheta = np.sum(np.square(sintheta_histo - sintheta_ref) /
+                             sintheta_ref) / sintheta_histo.size
+
+    print('chisqr_phi = ',chisqr_phi)
+    print('chisqr_sintheta = ',chisqr_sintheta)
+    assert 0.9 < chisqr_phi < 1.1, "Distribution in azimuth is not nearly uniform"
+    assert 0.9 < chisqr_sintheta < 1.1, "Distribution in sin(inclination) is not nearly uniform"
+
+    # Try some invalid inputs
+    try:
+        np.testing.assert_raises(ValueError, galsim.FRatioAngles, fratio=-0.3)
+        np.testing.assert_raises(ValueError, galsim.FRatioAngles, fratio=1.2, obscuration=-0.3)
+        np.testing.assert_raises(ValueError, galsim.FRatioAngles, fratio=1.2, obscuration=1.0)
+        np.testing.assert_raises(ValueError, galsim.FRatioAngles, fratio=1.2, obscuration=1.9)
+    except ImportError:
+        pass
+
 
 if __name__ == '__main__':
     test_photon_array()
     test_wavelength_sampler()
+    test_photon_angles()